pyproject.toml
[tool.poetry]
name = "enigma-auth-system"
version = "0.1.0"
description = "Enigma authentication API"
authors = ["Koech"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
